# docs/openapi.yaml
openapi: 3.1.1
info:
  title: API Lanchonete
  version: 1.0.0
  description: >
    API REST para gerenciamento de pedidos,
    produtos, estoque, usuários, pagamentos
    e unidades da lanchonete.
servers:
  - url: https://lanchonete-api.onrender.com
    description: Servidor de produção

  - url: http://127.0.0.1:5000
    description: Servidor local
...
paths:
...
  /login/pedidos:
    post:
      tags:
      - Pedidos
      summary: Realizar novo pedido
      description: Permite que o cliente realize um pedido, desde que esteja autenticado.
      operationId: novoPedido
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - unidadeId
              - entrega
              - canalPedido
              - usarPontos
              - itempedido
              properties:
                unidadeId:
                  type: integer
                  description: ID da unidade em que será feito o pedido.
                  example: 2
                observacao:
                  type: string
                  description: Descrição do pedido.
                  example: "Retirar cebola."
                entrega:
                  type: boolean
                  description: Informa se o pedido terá ou não entrega.
                  example: true
                usarPontos:
                  type: boolean
                  description: Define se os pontos do cliente serão utilizados ou não. O uso só é permitido quando 
                    houver saldo mínimo suficiente de pontos.
                  example: true
                canalPedido:
                  type: string
                  enum:
                  - TOTEM
                  - WEB
                  - BALCAO
                  - PICKUP
                  - APP
                  description: Canal em que o pedido foi realizado.
                  example: "TOTEM"
                itempedido:
                  type: array
                  minItems: 1
                  description: Itens que compõem o pedido.
                  items:
                    type: object
                    required:
                    - produtoId
                    - quantidade
                    properties:
                      produtoId:
                        type: integer
                        minimum: 1
                        example: 3
                      quantidade:
                        type: integer
                        minimum: 1
                        example: 2
      responses:
        "201":
          description: Pedido realizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespostaSucesso'
              example:
                message: "Pedido realizado com sucesso."
                data:
                  id: 2
                  unidadeId: 4
                  status: AGUARDANDO_PAGAMENTO
                  total: 35.6
                  dataPedido: "2026-08-06 23:57:22"
                  entrega: false
                  usarPontos: true
                  observacao: "Retirar cebola."
                  canalPedido: BALCAO
                  itempedido: [{estoqueId: 3, quantidade: 2, preco: 17.8, subtotal: 35.6}]
        "400":
          description: Requisição inválida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "JSON_INVALIDO"
                message: "O corpo da requisição deve ser um JSON válido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "409":
          description: Conflito com o estado atual do recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Dados inválidos.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "CANAL_INVALIDO"
                message: "O canal informado não existe."
                details: [{field: "canalPedido", issue: "Valores aceitos: TOTEM, WEB, PICKUP, BALCAO, APP"}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/login/pedidos"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        '401':
          description: Token JWT ausente, expirado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Usuário autenticado sem permissão para acessar este recurso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /{pedidoId}/status/aceitar:
    patch:
      tags:
      - Pedidos
      summary: Atualizar o status do pedido.
      description: >
        Após feito o pagamento, o balconista pode aceitar o pedido para poder iniciar o preparo. Apenas cargos de gestão e
        o balconista podem realizar esta função.
      operationId: aceitarPedido
      security:
      - bearerAuth: []
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: O pedido que terá seu status alterado.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Pedido aceito com sucesso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Pedido aceito e enviado para preparo."
                data:
                  pedidoId: 2
                  status: EM_PREPARO
                  unidadeId: 1
                  usuarioId: 3
                  dataPedido: "2026-08-06 23:57:22"
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/aceitar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/aceitar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PEDIDO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este pedido."
                details: [{field: "pedidoId", issue: "O pedido com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/aceitar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/aceitar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /{pedidoId}/status/pronto:
    patch:
      tags:
      - Pedidos
      summary: Atualizar o status do pedido.
      description: >
        Após o preparo iniciado e o produto já feito, o cozinheiro deve alterar o status para que avise ao cliente. Apenas
        cargos de gestão e o cozinheiro podem realizar esta função.
      operationId: pedidoPronto
      security:
      - bearerAuth: []
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: O pedido que terá seu status alterado.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Pedido já preparado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Pedido pronto."
                data:
                  pedidoId: 2
                  status: PRONTO
                  unidadeId: 1
                  usuarioId: 3
                  dataPedido: "2026-08-06 23:57:22"
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/pronto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/pronto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PEDIDO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este pedido."
                details: [{field: "pedidoId", issue: "O pedido com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/pronto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/pronto"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /{pedidoId}/status/entrega:
    patch:
      tags:
      - Pedidos
      summary: Atualizar o status do pedido.
      description: >
        Após pronto, caso o pedido seja para entregar, o balconista deve alterar o status quando o entregador sair da loja.
        Apenas cargos de gestão e o atendente podem realizar esta função.
      operationId: entregarPedido
      security:
      - bearerAuth: []
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: O pedido que terá seu status alterado.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Pedido saindo para entrega.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Pedido saiu para entrega."
                data:
                  pedidoId: 2
                  status: AGUARDANDO_ENTREGADOR
                  unidadeId: 1
                  usuarioId: 3
                  dataPedido: "2026-08-06 23:57:22"
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/entrega"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/entrega"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PEDIDO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este pedido."
                details: [{field: "pedidoId", issue: "O pedido com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/entrega"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/entrega"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /{pedidoId}/status/finalizar:
    patch:
      tags:
      - Pedidos
      summary: Atualizar o status do pedido.
      description: >
        Após pedido entregue, seja pelo motoboy ou para o cliente diretamente, registrar como finalizado para poder finalizar.
        Apenas cargos de gestão e o atendente podem realizar esta função.
      operationId: finalizarPedido
      security:
      - bearerAuth: []
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: O pedido que terá seu status alterado.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Pedido finalizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Pedido finalizado com sucesso."
                data:
                  pedidoId: 2
                  status: FINALIZADO
                  unidadeId: 1
                  usuarioId: 3
                  dataPedido: "2026-08-06 23:57:22"
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/finalizar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/finalizar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PEDIDO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este pedido."
                details: [{field: "pedidoId", issue: "O pedido com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/finalizar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/finalizar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
  /{pedidoId}/status/cancelar:
    patch:
      tags:
      - Pedidos
      summary: Atualizar o status do pedido.
      description: >
        Caso precise cancelar o pedido. Apenas cargos de gestão, atendente e clientes podem realizar esta função.
      operationId: cancelarPedido
      security:
      - bearerAuth: []
      parameters:
      - name: pedidoId
        in: path
        required: true
        description: O pedido que terá seu status alterado.
        schema:
          type: integer
          minimum: 1
      responses:
        "200":
          description: Pedido cancelado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaSucesso"
              example:
                message: "Pedido cancelado com sucesso."
                data:
                  pedidoId: 2
                  status: CANCELADO
                  unidadeId: 1
                  usuarioId: 3
                  dataPedido: "2026-08-06 23:57:22"
        "401":
          description: Token JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "TOKEN_ACESSO"
                message: "Token de acesso expirado ou inválido."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/cancelar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "403":
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "USUARIO_NAO_PERMITIDO"
                message: "O usuário logado não tem permissão para executar esta ação."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/cancelar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "404":
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "PEDIDO_NAO_ENCONTRADO"
                message: "Não foi possível encontrar este pedido."
                details: [{field: "pedidoId", issue: "O pedido com o ID 2 não existe."}]
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/cancelar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
        "500":
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              example:
                error: "ERRO_INTERNO"
                message: "Ocorreu um erro interno no servidor."
                details: []
                timestamp: "2026-08-07T08:30:00-03:00"
                path: "/2/status/cancelar"
                requestId: "baaf6e1b-5217-41af-9877-2298562e65e3"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          example: "USUARIO_NAO_ENCONTRADO"
        message:
          type: string
          example: "O usuário informado não foi encontrado."
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "usuario"
              issue:
                type: string
                example: "O usuário informado não existe."
        timestamp:
          type: string
          format: date-time
          example: "2026-08-07T08:30:00-03:00"
        path:
          type: string
          example: "/login"
        requestId:
          type: string
          format: uuid
          example: "baaf6e1b-5217-41af-9877-2298562e65e3"
...
    RespostaSucesso:
      type: object
      properties:
        message:
          type: string
          example: Operação realizada com sucesso.
        data: {}
...